pyproject.toml
[tool.poetry]
name = 'dotrun'
version = '2.0.1'
description = 'A tool for developing Node.js and Python projects'
authors = ['Canonical Web Team <webteam@canonical.com>']
license = 'LGPL-3.0'
readme = 'README.md'

[tool.poetry.scripts]
dotrun = "dotrun:cli"

[tool.poetry.dependencies]
python = '^3.6'
python-dotenv = '0.20.0'
python-slugify = '6.1.2'
docker = '5.0.3'
dockerpty = '0.4.1'


[build-system]
requires = ['poetry>=0.12']
build-backend = 'poetry.masonry.api'
